fix: package and folder names
