+++ official/docs/java/carrier_accounts/DeleteCarrrierAccount.java
@@ -1,14 +1,15 @@
-package carrierAccounts;
+package carrier_accounts;
 
 import com.easypost.EasyPost;
 import com.easypost.exception.EasyPostException;
 import com.easypost.model.CarrierAccount;
 
-public class RetrieveCarrierAccount {
+public class DeleteCarrrierAccount {
     public static void main(String[] args) throws EasyPostException {
         EasyPost.apiKey = System.getenv("EASYPOST_API_KEY");
 
         CarrierAccount carrierAccount = CarrierAccount.retrieve("ca_...");
+        carrierAccount.delete();
 
         System.out.println(carrierAccount);
     }
