+++ official/docs/java/carrier_accounts/RetrieveCarrierAccount.java
@@ -1,17 +1,15 @@
-package carrierAccounts;
+package carrier_accounts;
 
 import com.easypost.EasyPost;
 import com.easypost.exception.EasyPostException;
 import com.easypost.model.CarrierAccount;
 
-import java.util.List;
-
-public class RetrieveCarrierAccountList {
+public class RetrieveCarrierAccount {
     public static void main(String[] args) throws EasyPostException {
         EasyPost.apiKey = System.getenv("EASYPOST_API_KEY");
 
-        List<CarrierAccount> carrierAccounts = CarrierAccount.all();
+        CarrierAccount carrierAccount = CarrierAccount.retrieve("ca_...");
 
-        System.out.println(carrierAccounts);
+        System.out.println(carrierAccount);
     }
 }
