+++ official/docs/java/carrier_accounts/RetrieveCarrierAccountList.java
@@ -1,16 +1,17 @@
-package carrierAccounts;
+package carrier_accounts;
 
 import com.easypost.EasyPost;
 import com.easypost.exception.EasyPostException;
 import com.easypost.model.CarrierAccount;
 
-public class DeleteCarrrierAccount {
+import java.util.List;
+
+public class RetrieveCarrierAccountList {
     public static void main(String[] args) throws EasyPostException {
         EasyPost.apiKey = System.getenv("EASYPOST_API_KEY");
 
-        CarrierAccount carrierAccount = CarrierAccount.retrieve("ca_...");
-        carrierAccount.delete();
+        List<CarrierAccount> carrierAccounts = CarrierAccount.all();
 
-        System.out.println(carrierAccount);
+        System.out.println(carrierAccounts);
     }
 }
